Move common internal stuff to Test.DejaFu.Common
