+++ dejafu/Test/DejaFu/Deterministic/Internal/Threading.hs
@@ -17,6 +17,8 @@ import Control.Exception (Exception, MaskingState(..), SomeException, fromExcept
 import Data.List (intersect, nub)
 import Data.Map.Strict (Map)
 import Data.Maybe (fromMaybe, isJust, isNothing)
+
+import Test.DejaFu.Common
 import Test.DejaFu.Deterministic.Internal.Common
 
 import qualified Data.Map.Strict as M
