+++ dejafu/Test/DejaFu/STM/Internal.hs
@@ -19,7 +19,8 @@ import Control.Exception (Exception, SomeException, fromException, toException)
 import Control.Monad.Cont (Cont, runCont)
 import Control.Monad.Ref (MonadRef, newRef, readRef, writeRef)
 import Data.List (nub)
-import Test.DejaFu.Deterministic.Internal.Common (TVarId, IdSource, TAction(..), TTrace, nextTVId)
+
+import Test.DejaFu.Common
 
 --------------------------------------------------------------------------------
 -- The @STMLike@ monad
